Tsx conversion, added to app/enter/display
