+++ pages/_app.tsx
@@ -4,6 +4,7 @@ import { SessionProvider } from 'next-auth/react';
 import Layout from '@/components/layout';
 import '@/styles/custom.scss'
 import { LoginProvider } from '@/context/LoginContext';
+import { GameProvider } from '@/context/GameContext';
 
 export default function MyApp({
   Component,
@@ -12,9 +13,11 @@ export default function MyApp({
   return (
     // <SessionProvider session={session}>
     <LoginProvider>
-      <Layout {...pageProps}>
-        <Component {...pageProps} />
-      </Layout>
+      <GameProvider>
+        <Layout {...pageProps}>
+          <Component {...pageProps} />
+        </Layout>
+      </GameProvider>
     </LoginProvider>
     // </SessionProvider>
   );
